@@ -1,27 +1,21 @@
 {
   "name": "client",
-  "proxy": "http://localhost:4000",
+  "proxy": "http://localhost:3000",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@emotion/react": "^11.11.0",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.13.2",
-    "@mui/styles": "^5.13.2",
-    "@p5-wrapper/react": "^4.1.2",
-    "bootstrap": "^5.2.3",
-    "react": "^18.0.0",
-    "react-bootstrap": "^2.7.4",
-    "react-dom": "^18.0.0",
-    "react-icons": "^4.9.0",
-    "react-lottie": "^1.2.3",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
-    "react-typist": "^2.0.5",
-    "react-youtube": "^10.1.0",
-    "styled-components": "^5.3.11",
-    "three": "^0.152.2"
+    "@p5-wrapper/react": "4.1.2",
+    "lottie-react": "^2.4.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "2.7.4",
+    "react-dom": "^18.2.0",
+    "react-icons": "4.9.0",
+    "react-router-dom": "^6.11.2",
+    "react-scripts": "^5.0.1",
+    "styled-components": "5.3.11",
+    "three": "^0.153.0",
+    "typed.js": "^2.0.16"
   },
   "resolutions": {
     "react-error-overlay": "6.0.9"
